core/src/test/java/hello/core/order/OrderServiceImplTest.java
package hello.core.order;

import hello.core.discount.FixDiscountPolicy;
import hello.core.member.Grade;
import hello.core.member.Member;
import hello.core.member.MemoryMemberRepository;

import org.assertj.core.api.Assertions;
import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.*;
import static org.junit.jupiter.api.Assertions.*;

class OrderServiceImplTest {
    //애는 순수한 자바로 테스트를 만들어 볼것임

    @Test
    void createOrder(){

        MemoryMemberRepository memberRepository = new MemoryMemberRepository();
        memberRepository.save(new Member(1L,"name", Grade.VIP));

        //테스트 짜는 입장에서는 이게 의존관계가 바로 안보임, 코드 까봐야 앎
        OrderServiceImpl orderService = new OrderServiceImpl(new MemoryMemberRepository(),new FixDiscountPolicy());
        //지금 가짜 객체 그냥 넣어줌
        Order order = orderService.createOrder(1L,"itemA",10000);
        assertThat(order.getDiscountPrice()).isEqualTo(1000);
    }
    //결론은 생성자 주입으로 해야 테스트 할때도 바로 뭐가 잘못됐는지 찾을 수 있다~

}